feat(frontend): extract component to load Ethereum transactions
