+++ src/frontend/src/eth/components/transactions/EthTransactions.svelte
@@ -5,10 +5,8 @@
 	import EthTransaction from '$eth/components/transactions/EthTransaction.svelte';
 	import EthTransactionModal from '$eth/components/transactions/EthTransactionModal.svelte';
 	import EthTransactionsSkeletons from '$eth/components/transactions/EthTransactionsSkeletons.svelte';
-	import { tokenNotInitialized } from '$eth/derived/nav.derived';
 	import { ethereumTokenId, ethereumToken } from '$eth/derived/token.derived';
 	import { sortedTransactions } from '$eth/derived/transactions.derived';
-	import { loadTransactions } from '$eth/services/transactions.services';
 	import type { EthTransactionUi } from '$eth/types/eth-transaction';
 	import { mapTransactionUi } from '$eth/utils/transactions.utils';
 	import { ckEthMinterInfoStore } from '$icp-eth/stores/cketh.store';
@@ -18,13 +16,10 @@
 	import { SLIDE_DURATION } from '$lib/constants/transition.constants';
 	import { ethAddress } from '$lib/derived/address.derived';
 	import { modalToken, modalTransaction } from '$lib/derived/modal.derived';
-	import { tokenWithFallback } from '$lib/derived/token.derived';
 	import { i18n } from '$lib/stores/i18n.store';
 	import { modalStore } from '$lib/stores/modal.store';
 	import type { OptionEthAddress } from '$lib/types/address';
-	import type { TokenId } from '$lib/types/token';
 	import type { Transaction as TransactionType } from '$lib/types/transaction';
-	import { isNetworkIdEthereum } from '$lib/utils/network.utils';
 
 	let ckMinterInfoAddresses: OptionEthAddress[] = [];
 	$: ckMinterInfoAddresses = toCkMinterInfoAddresses({
@@ -40,42 +35,6 @@
 			$ethAddress: $ethAddress
 		})
 	);
-
-	let tokenIdLoaded: TokenId | undefined = undefined;
-
-	const load = async () => {
-		if ($tokenNotInitialized) {
-			tokenIdLoaded = undefined;
-			return;
-		}
-
-		const {
-			network: { id: networkId },
-			id: tokenId
-		} = $tokenWithFallback;
-
-		// If user browser ICP transactions but switch token to Eth, due to the derived stores, the token can briefly be set to ICP while the navigation is not over.
-		// This prevents the glitch load of ETH transaction with a token ID for ICP.
-		if (!isNetworkIdEthereum(networkId)) {
-			tokenIdLoaded = undefined;
-			return;
-		}
-
-		// We don't reload the same token in a row.
-		if (tokenIdLoaded === tokenId) {
-			return;
-		}
-
-		tokenIdLoaded = tokenId;
-
-		const { success } = await loadTransactions({ tokenId, networkId });
-
-		if (!success) {
-			tokenIdLoaded = undefined;
-		}
-	};
-
-	$: $tokenWithFallback, $tokenNotInitialized, (async () => await load())();
 
 	let selectedTransaction: TransactionType | undefined;
 	$: selectedTransaction = $modalTransaction
