+++ src/frontend/src/routes/(app)/transactions/+layout.svelte
@@ -2,8 +2,11 @@
 	import Listener from '$lib/components/core/Listener.svelte';
 
 	import { token } from '$lib/stores/token.store';
+	import LoaderTransactions from '$lib/components/loaders/LoaderTransactions.svelte';
 </script>
 
 <Listener token={$token}>
-	<slot />
+	<LoaderTransactions>
+		<slot />
+	</LoaderTransactions>
 </Listener>
